backend/requirements.txt
# Django Core
Django
djangorestframework
django-cors-headers
python-decouple

# Database
psycopg2-binary
dj-database-url

# Authentication & Security
PyJWT
cryptography
django-encrypted-model-fields

# GitHub Integration
requests==2.31.0
PyGithub==2.1.1

# AI Integration
google-genai==0.2.2

# API Documentation
drf-spectacular==0.27.1
# Background Tasks
celery
redis

# Utilities
python-dotenv
Pillow

# Development
django-extensions
ipython
